+++ core/database/src/main/java/com/andanana/musicplayer/core/database/usecases/GetMusicInPlayList.kt
@@ -21,7 +21,7 @@ class GetMusicInPlayList @Inject constructor(
                 musics.map { music ->
                     val crossRef = getPlayListMusicCrossRef(
                         playList.playListId,
-                        musicId = music.id
+                        musicId = music.id.toLong()
                     )
                     MusicWithAddedTime(
                         musicEntity = music,
